requirements.txt
fastapiuvicornyfinancepandaspandas_ta